requirements.txt
# Python 3.10+
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
torch>=2.0.0
PyYAML>=6.0
tqdm>=4.65.0
streamlit>=1.54.0
